switch interface and move types inside of authorization module
