+++ apps/server/src/modules/authorization/authorization.service.ts
@@ -1,20 +1,14 @@
-import { ForbiddenException, Injectable, UnauthorizedException } from '@nestjs/common';
-import { BaseDO, EntityId, User } from '@shared/domain';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { BaseDO, User } from '@shared/domain';
 import { AuthorizableObject } from '@shared/domain/domain-object';
-import { ErrorUtils } from '@src/core/error/utils';
 import { AuthorizationHelper } from './authorization.helper';
 import { ForbiddenLoggableException } from './errors/forbidden.loggable-exception';
-import { ReferenceLoader } from './domain/reference/reference.loader';
 import { RuleManager } from './rule-manager';
-import { AuthorizableReferenceType, AuthorizationContext } from './types';
+import { AuthorizationContext } from './types';
 
 @Injectable()
 export class AuthorizationService {
-	constructor(
-		private readonly ruleManager: RuleManager,
-		private readonly loader: ReferenceLoader,
-		private readonly authorizationHelper: AuthorizationHelper
-	) {}
+	constructor(private readonly ruleManager: RuleManager, private readonly authorizationHelper: AuthorizationHelper) {}
 
 	public checkPermission(user: User, object: AuthorizableObject | BaseDO, context: AuthorizationContext): void {
 		if (!this.hasPermission(user, object, context)) {
@@ -27,47 +21,6 @@ export class AuthorizationService {
 		const hasPermission = rule.hasPermission(user, object, context);
 
 		return hasPermission;
-	}
-
-	/**
-	 * @deprecated
-	 */
-	public async checkPermissionByReferences(
-		userId: EntityId,
-		entityName: AuthorizableReferenceType,
-		entityId: EntityId,
-		context: AuthorizationContext
-	): Promise<void> {
-		if (!(await this.hasPermissionByReferences(userId, entityName, entityId, context))) {
-			throw new ForbiddenLoggableException(userId, entityName, context);
-		}
-	}
-
-	/**
-	 * @deprecated
-	 */
-	public async hasPermissionByReferences(
-		userId: EntityId,
-		entityName: AuthorizableReferenceType,
-		entityId: EntityId,
-		context: AuthorizationContext
-	): Promise<boolean> {
-		// TODO: This try-catch-block should be removed. See ticket: https://ticketsystem.dbildungscloud.de/browse/BC-4023
-		try {
-			const [user, object] = await Promise.all([
-				this.getUserWithPermissions(userId),
-				this.loader.loadAuthorizableObject(entityName, entityId),
-			]);
-			const rule = this.ruleManager.selectRule(user, object, context);
-			const hasPermission = rule.hasPermission(user, object, context);
-
-			return hasPermission;
-		} catch (error) {
-			throw new ForbiddenException(
-				null,
-				ErrorUtils.createHttpExceptionOptions(error, 'AuthorizationService:hasPermissionByReferences')
-			);
-		}
 	}
 
 	public checkAllPermissions(user: User, requiredPermissions: string[]): void {
@@ -91,10 +44,11 @@ export class AuthorizationService {
 	public hasOneOfPermissions(user: User, requiredPermissions: string[]): boolean {
 		return this.authorizationHelper.hasOneOfPermissions(user, requiredPermissions);
 	}
-
+	/* think about it
 	public async getUserWithPermissions(userId: EntityId): Promise<User> {
 		const userWithPermissions = await this.loader.getUserWithPermissions(userId);
 
 		return userWithPermissions;
 	}
+*/
 }
