+++ apps/server/src/modules/authorization/domain/reference/authorization-reference.service.spec.ts
@@ -1,9 +1,9 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
-import { ForbiddenException, NotFoundException } from '@nestjs/common';
+import { NotFoundException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { courseFactory, setupEntities, userFactory } from '@shared/testing';
 import { ObjectId } from 'bson';
-import { AuthorizableReferenceType } from '../../types';
+import { AuthorizableReferenceType } from './types';
 import { AuthorizationService } from '../../authorization.service';
 import { ReferenceLoader } from './reference.loader';
 import { AuthorizationContextBuilder } from '../../authorization-context.builder';
