+++ apps/server/src/modules/authorization/domain/reference/types/allowed-authorization-object-type.enum.ts
