+++ apps/server/src/modules/authorization/domain/reference/types/index.ts
@@ -1,0 +1,2 @@
+export * from './authorization-loader-service';
+export * from './allowed-authorization-object-type.enum';
