+++ apps/server/src/modules/authorization/types/index.ts
@@ -1,5 +1,3 @@
 export * from './action.enum';
 export * from './authorization-context.interface';
 export * from './rule.interface';
-export * from './allowed-authorization-object-type.enum';
-export * from './authorization-loader-service';
